曲一つのプロパティ情報を更新するバックエンド API を実装

+++ backend/src/track_property/single_track_property.rs
@@ -46,7 +46,7 @@ pub struct SingleTrackProperty {
     pub release_date: Option<NaiveDate>,
 
     /// レート (好み)
-    pub rating: i32,
+    pub rating: i16,
     /// 原曲
     pub original_track: String,
     /// メモ
